Move getConfirmedRequests from event-service to request-service

+++ core/interaction-api/src/main/java/ru/practicum/explore/with/me/interaction/api/dto/event/EventRequestCount.java
@@ -1,4 +1,4 @@
 package ru.practicum.explore.with.me.interaction.api.dto.event;
 
-public record EventRequestCount(Long eventId, Long count) {
+public record EventRequestCount(Long eventId, Integer count) {
 }
